fix prosemirror-model version conflict
