+++ package.json
@@ -24,6 +24,9 @@
     "vuex": "^3.4.0",
     "vuex-module-decorators": "^0.17.0"
   },
+  "resolutions": {
+    "prosemirror-model": "1.8.2"
+  },
   "devDependencies": {
     "@types/chai": "^4.2.11",
     "@types/mocha": "^7.0.2",
