+++ src/views/Post/Layout.vue
@@ -10,8 +10,5 @@ import { Component, Vue } from 'vue-property-decorator';
   components: {},
 })
 export default class extends Vue {
-  private created() {
-    console.log('layout');
-  }
 }
 </script>
